end sprint 3 without tests and checks

* end sprint 3 without tests and checks

* Merge branch 'main' into change-permission-check

+++ src/main/java/superapp/service/impl/AdminServiceImpl.java
@@ -2,30 +2,32 @@ package superapp.service.impl;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.core.env.Environment;
 import org.springframework.stereotype.Service;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
-import superapp.boundary.MiniAppCommandBoundary;
-import superapp.boundary.UserBoundary;
+import superapp.boundary.command.MiniAppCommandBoundary;
+import superapp.boundary.user.UserBoundary;
 
+
+import superapp.entity.user.UserId;
+import superapp.entity.user.UserRole;
 import superapp.repository.MiniAppCommandsRepository;
 import superapp.repository.ObjectRepository;
 import superapp.repository.UserRepository;
+import superapp.service.AbstractService;
 import superapp.service.AdminService;
+
+import static superapp.exception.Consts.ADMIN_PERMISSION_EXCEPTION;
 
 
 @Service
-public class AdminServiceImpl implements AdminService {
+public class AdminServiceImpl extends AbstractService implements AdminService {
 
     private static final Logger logger = LoggerFactory.getLogger(AdminServiceImpl.class);
 
     private final UserRepository userRepository;
     private final ObjectRepository objectRepository;
     private final MiniAppCommandsRepository miniAppCommandsRepository;
-    @Autowired
-    private Environment environment;
 
     public AdminServiceImpl(UserRepository userRepository,
                             MiniAppCommandsRepository miniAppCommandsRepository,
@@ -38,41 +40,81 @@ public class AdminServiceImpl implements AdminService {
 
 
     @Override
-    public Mono<Void> deleteAllUsers() {
+    public Mono<Void> deleteAllUsers(String superApp, String email) {
         logger.info("Deleting all users in AdminServiceImpl");
-        return userRepository.deleteAll().log();
+        UserId userId = new UserId(superApp, email);
+        return isValidUserCredentials(userId, UserRole.ADMIN, userRepository)
+                .flatMap(isValid -> {
+                    if (!isValid) {
+                        return Mono.error(new IllegalAccessException(ADMIN_PERMISSION_EXCEPTION));
+                    }
+                    return userRepository.deleteAll().log();
+                });
     }
 
     @Override
-    public Mono<Void> deleteAllObjects() {
+    public Mono<Void> deleteAllObjects(String superApp, String email) {
         logger.info("Deleting all objects in AdminServiceImpl");
-        return objectRepository.deleteAll().log();
+        UserId userId = new UserId(superApp, email);
+        return isValidUserCredentials(userId, UserRole.ADMIN, userRepository)
+                .flatMap(isValid -> {
+                    if (!isValid) {
+                        return Mono.error(new IllegalAccessException(ADMIN_PERMISSION_EXCEPTION));
+                    }
+                    return objectRepository.deleteAll().log();
+                });
     }
 
     @Override
-    public Mono<Void> deleteAllCommandsHistory() {
+    public Mono<Void> deleteAllCommandsHistory(String superApp, String email) {
         logger.info("Deleting all commands history in AdminServiceImpl");
-        return miniAppCommandsRepository.deleteAll().log();
+        UserId userId = new UserId(superApp, email);
+        return isValidUserCredentials(userId, UserRole.ADMIN, userRepository)
+                .flatMap(isValid -> {
+                    if (!isValid) {
+                        return Mono.error(new IllegalAccessException(ADMIN_PERMISSION_EXCEPTION));
+                    }
+                    return miniAppCommandsRepository.deleteAll().log();
+                });
     }
 
     @Override
-    public Flux<UserBoundary> exportAllUsers() {
+    public Flux<UserBoundary> exportAllUsers(String superApp, String email) {
         logger.info("Exporting all users");
-        return userRepository.findAll().map(UserBoundary::new).log();
+        UserId userId = new UserId(superApp, email);
+        return isValidUserCredentials(userId, UserRole.ADMIN, userRepository)
+                .flatMapMany(isValid -> {
+                    if (!isValid) {
+                        return Flux.error(new IllegalAccessException(ADMIN_PERMISSION_EXCEPTION));
+                    }
+                    return userRepository.findAll().map(UserBoundary::new).log();
+                });
     }
 
     @Override
-    public Flux<MiniAppCommandBoundary> exportAllMiniAppsCommandsHistory() {
+    public Flux<MiniAppCommandBoundary> exportAllMiniAppsCommandsHistory(String superApp, String email) {
         logger.info("Exporting all Miniapps commands history");
-        return miniAppCommandsRepository.findAll().map(MiniAppCommandBoundary::new).log();
+        UserId userId = new UserId(superApp, email);
+        return isValidUserCredentials(userId, UserRole.ADMIN, userRepository)
+                .flatMapMany(isValid -> {
+                    if (!isValid) {
+                        return Flux.error(new IllegalAccessException(ADMIN_PERMISSION_EXCEPTION));
+                    }
+                    return miniAppCommandsRepository.findAll().map(MiniAppCommandBoundary::new).log();
+                });
     }
 
 
     @Override
-    public Flux<MiniAppCommandBoundary> exportMiniAppCommandsHistory(String miniAppName) {
+    public Flux<MiniAppCommandBoundary> exportMiniAppCommandsHistory(String miniAppName, String superApp, String email) {
         logger.info("Exporting history of miniapp: " + miniAppName);
-        return miniAppCommandsRepository
-                .findByCommandIdMiniApp(miniAppName)
-                .map(MiniAppCommandBoundary::new).log();
+        UserId userId = new UserId(superApp, email);
+        return isValidUserCredentials(userId, UserRole.ADMIN, userRepository)
+                .flatMapMany(isValid -> {
+                    if (!Boolean.TRUE.equals(isValid)) {
+                        return Flux.error(new IllegalAccessException(ADMIN_PERMISSION_EXCEPTION));
+                    }
+                    return miniAppCommandsRepository.findByCommandIdMiniApp(miniAppName).map(MiniAppCommandBoundary::new).log();
+                });
     }
 }
